chore: rename fs-utils -> fs, string-utils -> string
